+++ apps/ledger-live-mobile/src/screens/AddAccounts/03-Accounts.js
@@ -24,10 +24,12 @@ import uniq from "lodash/uniq";
 import { Trans } from "react-i18next";
 import type {
   CryptoCurrency,
+  TokenCurrency,
   Account,
 } from "@ledgerhq/live-common/types/index";
 import { getCurrencyBridge } from "@ledgerhq/live-common/bridge/index";
 import type { Device } from "@ledgerhq/live-common/hw/actions/types";
+import { isTokenCurrency } from "@ledgerhq/live-common/currencies/index";
 
 import type { DerivationMode } from "@ledgerhq/live-common/derivation";
 
@@ -68,7 +70,7 @@ const SectionAccounts = ({ defaultSelected, ...rest }: any) => {
 };
 
 type RouteParams = {
-  currency: CryptoCurrency,
+  currency: CryptoCurrency | TokenCurrency,
   device: Device,
   inline?: boolean,
   returnToSwap?: boolean,
@@ -182,7 +184,10 @@ function AddAccountsAccounts({
   }, []);
 
   const startSubscription = useCallback(() => {
-    const bridge = getCurrencyBridge(currency);
+    const cryptoCurrency = isTokenCurrency(currency)
+      ? currency.parentCurrency
+      : currency;
+    const bridge = getCurrencyBridge(cryptoCurrency);
     const syncConfig = {
       paginationConfig: {
         operation: 0,
@@ -192,9 +197,9 @@ function AddAccountsAccounts({
     // will be set to false if an existing account is found
 
     scanSubscription.current = concat(
-      from(prepareCurrency(currency)).pipe(ignoreElements()),
+      from(prepareCurrency(cryptoCurrency)).pipe(ignoreElements()),
       bridge.scanAccounts({
-        currency,
+        currency: cryptoCurrency,
         deviceId,
         syncConfig,
       }),