+++ libs/ledger-live-common/src/platform/converters.ts
@@ -1,13 +1,44 @@
 import byFamily from "../generated/platformAdapter";
-
-import type { Account, CryptoCurrency, Transaction } from "../types";
-import type {
+import { isTokenAccount, isSubAccount } from "../account";
+import { Account, AccountLike, Transaction } from "../types";
+import {
   PlatformAccount,
   PlatformCurrency,
   PlatformTransaction,
+  PlatformCurrencyType,
+  PlatformTokenStandard,
+  PlatformSupportedCurrency,
 } from "./types";
 
-export function accountToPlatformAccount(account: Account): PlatformAccount {
+export function accountToPlatformAccount(
+  account: AccountLike,
+  parentAccount?: Account
+): PlatformAccount {
+  if (isSubAccount(account)) {
+    if (!parentAccount) {
+      throw new Error("No 'parentAccount' account provided for token account");
+    }
+
+    return {
+      id: account.id,
+      balance: account.balance,
+      address: parentAccount.freshAddress,
+      blockHeight: parentAccount.blockHeight,
+      lastSyncDate: parentAccount.lastSyncDate,
+      ...(isTokenAccount(account)
+        ? {
+            name: `${parentAccount.name} (${account.token.ticker})`,
+            currency: account.token.id,
+            spendableBalance: account.spendableBalance,
+          }
+        : {
+            name: account.name,
+            currency: account.currency.id,
+            spendableBalance: parentAccount.spendableBalance,
+          }),
+    };
+  }
+
   return {
     id: account.id,
     name: account.name,
@@ -19,11 +50,34 @@ export function accountToPlatformAccount(account: Account): PlatformAccount {
     lastSyncDate: account.lastSyncDate,
   };
 }
+
 export function currencyToPlatformCurrency(
-  currency: CryptoCurrency
+  currency: PlatformSupportedCurrency
 ): PlatformCurrency {
+  if (currency.type === "TokenCurrency") {
+    if (currency.parentCurrency.family !== "ethereum") {
+      throw new Error("Only ERC20 tokens are supported");
+    }
+
+    return {
+      type: PlatformCurrencyType.TokenCurrency,
+      standard: PlatformTokenStandard.ERC20,
+      id: currency.id,
+      ticker: currency.ticker,
+      contract: currency.contractAddress,
+      name: currency.name,
+      parent: currency.parentCurrency.id,
+      color: currency.parentCurrency.color,
+      units: currency.units.map((unit) => ({
+        name: unit.name,
+        code: unit.code,
+        magnitude: unit.magnitude,
+      })),
+    };
+  }
+
   return {
-    type: currency.type,
+    type: PlatformCurrencyType.CryptoCurrency,
     id: currency.id,
     ticker: currency.ticker,
     name: currency.name,